Add annotation to prevent uninstalling helm chart of Job/App (#231)
Sometimes we want to remove App/Job CR but keep their helm chart
running, it is a kind of soft self-uninstalling)

  To handle this use-case, ketch-controller looks for
theketch.io/dont-uninstall-helm-chart=true annotation.
If it is present on an App/Job CR and kuberentes has marked the CR for deletion,
ketch controller doesn't uninstall its helm chart.

+++ internal/controllers/job_controller.go
@@ -196,13 +196,15 @@ func (r *JobReconciler) deleteChart(ctx context.Context, job *ketchv1.Job) error
 			continue
 		}
 
-		helmClient, err := r.HelmFactoryFn(framework.Spec.NamespaceName)
-		if err != nil {
-			return err
-		}
-		err = helmClient.DeleteChart(job.Name)
-		if err != nil {
-			return err
+		if uninstallHelmChart(ketchv1.Group, job.Annotations) {
+			helmClient, err := r.HelmFactoryFn(framework.Spec.NamespaceName)
+			if err != nil {
+				return err
+			}
+			err = helmClient.DeleteChart(job.Name)
+			if err != nil {
+				return err
+			}
 		}
 		patchedFramework := framework
